+++ BooleanTasks/src/language/Node.java
@@ -1,6 +1,6 @@
 package language;
 public abstract class Node {
-	public final static float MIN_THICKNESS =3;
+	public final static double MIN_THICKNESS =3;
 	
 	//TODO: add error checkignt to model thickness, minimum priority
 	protected Node(){
