+++ BooleanTasks/src/language/Rotate.java
@@ -2,7 +2,7 @@ package language;
 
 public class Rotate extends Modifier {
 
-	public Rotate(Node n,float x, float y, float z) {
+	public Rotate(Node n,double x, double y, double z) {
 		super(n,"rotate(["+x+","+y+","+z+"]) ");
 	}
