+++ BooleanTasks/src/language/Translate.java
@@ -1,7 +1,7 @@
 package language;
 public class Translate extends Modifier {
 
-	public Translate(Node n,float x, float y, float z) {
+	public Translate(Node n,double x, double y, double z) {
 		super(n,"translate(["+x+","+y+","+z+"]) ");
 	}
