+++ BooleanTasks/src/language/TranslateSet.java
@@ -4,15 +4,15 @@ import java.util.ArrayList;
 
 public class TranslateSet extends Set{
 
-	public TranslateSet(ArrayList<Node> c, float x, float y, float z) {
+	public TranslateSet(ArrayList<Node> c, double x, double y, double z) {
 		super(c, "translate(["+x+","+y+","+z+"]) ");
 	}
 	
-	public TranslateSet(float x, float y, float z){
+	public TranslateSet(double x, double y, double z){
 		super("translate(["+x+","+y+","+z+"]) ");
 	}
 	
-	public TranslateSet(Set s, float x, float y, float z){
+	public TranslateSet(Set s, double x, double y, double z){
 		super(s.children,"translate(["+x+","+y+","+z+"]) " );
 	}
